+++ HandOfLesser/src/core/settings_global.cpp
