+++ HandOfLesser/src/core/settings_global.h
