+++ HandOfLesser/src/core/ui/display_global.cpp
