+++ HandOfLesser/src/hands/simple_gesture_detector.cpp
