+++ HandOfLesser/src/hands/simple_gesture_detector.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "openxr_hand.h"
 #include "simple_gesture.h"
+#include <src/openxr/openxr_hand.h>
 
 namespace HOL::SimpleGesture
 {
