+++ HandOfLesser/src/openxr/openxr_hand.cpp
@@ -1,17 +1,16 @@
 #include "openxr_hand.h"
+#include "XrUtils.h"
 #include <HandOfLesserCommon.h>
 #include "HandTrackingInterface.h"
-#include "settings_global.h"
-#include "display_global.h"
-#include "math_utils.h"
+#include "src/core/settings_global.h"
+#include "src/core/ui/display_global.h"
 #include <iostream>
 #include <utility>
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 
-using namespace HOL;
-using namespace HOL::SimpleGesture;
+using namespace HOL::OpenXR;
 
 OpenXRHand::OpenXRHand(xr::UniqueDynamicSession& session, XrHandEXT side)
 {
@@ -27,7 +26,7 @@ void OpenXRHand::init(xr::UniqueDynamicSession& session, XrHandEXT side)
 void OpenXRHand::calculateCurlSplay()
 {
 	const auto getJointOrientation = [&](XrHandJointEXT joint)
-	{ return HOL::toEigenQuaternion(this->mJointLocations[joint].pose.orientation); };
+	{ return toEigenQuaternion(this->mJointLocations[joint].pose.orientation); };
 
 	const auto getRootOpenXRJointForFinger = [&](FingerType joint)
 	{
@@ -104,15 +103,15 @@ void OpenXRHand::updateJointLocations(xr::UniqueDynamicSpace& space, XrTime time
 	auto palmLocation = this->mJointLocations[XrHandJointEXT::XR_HAND_JOINT_PALM_EXT];
 	auto palmVelocity = this->mJointVelocities[XrHandJointEXT::XR_HAND_JOINT_PALM_EXT];
 
-	this->handPose.palmLocation.position = HOL::toEigenVector(palmLocation.pose.position);
-	this->handPose.palmLocation.orientation = HOL::toEigenQuaternion(palmLocation.pose.orientation);
+	this->handPose.palmLocation.position = toEigenVector(palmLocation.pose.position);
+	this->handPose.palmLocation.orientation = toEigenQuaternion(palmLocation.pose.orientation);
 
 	// Orientation is not going to be set without position for hand tracking.
 	this->handPose.poseValid = (palmLocation.locationFlags & XR_SPACE_LOCATION_POSITION_VALID_BIT)
 							   == XR_SPACE_LOCATION_POSITION_VALID_BIT;
 
-	this->handPose.palmVelocity.linearVelocity = HOL::toEigenVector(palmVelocity.linearVelocity);
-	this->handPose.palmVelocity.angularVelocity = HOL::toEigenVector(palmVelocity.angularVelocity);
+	this->handPose.palmVelocity.linearVelocity = toEigenVector(palmVelocity.linearVelocity);
+	this->handPose.palmVelocity.angularVelocity = toEigenVector(palmVelocity.angularVelocity);
 
 	// Transform palm ( not actually same as grip ) to index controller tracker position
 	Eigen::Vector3f gripRotationOffset = Eigen::Vector3f(63, 0, -99);
@@ -136,7 +135,7 @@ void OpenXRHand::updateJointLocations(xr::UniqueDynamicSpace& space, XrTime time
 	// Rot
 	///////////
 
-	Eigen::Quaternionf rawOrientation = HOL::toEigenQuaternion(palmLocation.pose.orientation);
+	Eigen::Quaternionf rawOrientation = toEigenQuaternion(palmLocation.pose.orientation);
 	Eigen::Quaternionf rotationOffsetQuat
 		= HOL::quaternionFromEulerAnglesDegrees(mainRotationOffset);
 
@@ -146,7 +145,7 @@ void OpenXRHand::updateJointLocations(xr::UniqueDynamicSpace& space, XrTime time
 	// Trans
 	/////////////
 
-	Eigen::Vector3f rawPosition = HOL::toEigenVector(palmLocation.pose.position);
+	Eigen::Vector3f rawPosition = toEigenVector(palmLocation.pose.position);
 	Eigen::Vector3f rotationOffsetLocal
 		= this->handPose.palmLocation.orientation * mainTranslationOffset;
 	this->handPose.palmLocation.position = rawPosition + rotationOffsetLocal;
