+++ HandOfLesserCommon/include/HandOfLesserCommon.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "../src/transport/nativetransport.h"
-#include "../src/packet/nativepacket.h"
-#include "../src/hand.h"
-#include "../src/finger_bend.h"
-#include "../src/math/fingers.h"
-#include "../src/math/utils.h"
+#include "src/transport/nativetransport.h"
+#include "src/packet/nativepacket.h"
+#include "src/hand/hand.h"
+#include "src//hand/finger_bend.h"
+#include "src/math/fingers.h"
+#include "src/math/math_utils.h"
