+++ HandOfLesserCommon/src/hand/hand.h
