+++ HandOfLesserCommon/src/math/math_utils.cpp
@@ -1,25 +1,20 @@
 #include "math_utils.h"
+#include <numbers> 
 
-#ifndef M_PI
-#define M_PI 3.1415926535
-#endif
 
 namespace HOL
 {
 	Eigen::Quaternionf quaternionFromEulerAngles(float x, float y, float z)
 	{
-		return Eigen::Quaternionf(
-			Eigen::AngleAxisf(z, Eigen::Vector3f::UnitZ())
-			* Eigen::AngleAxisf(y, Eigen::Vector3f::UnitY())
-			* Eigen::AngleAxisf(x, Eigen::Vector3f::UnitX())
-		);
+		return Eigen::Quaternionf(Eigen::AngleAxisf(z, Eigen::Vector3f::UnitZ())
+								  * Eigen::AngleAxisf(y, Eigen::Vector3f::UnitY())
+								  * Eigen::AngleAxisf(x, Eigen::Vector3f::UnitX()));
 	}
 
 	Eigen::Quaternionf quaternionFromEulerAnglesDegrees(float x, float y, float z)
 	{
 		return quaternionFromEulerAngles(
-			degreesToRadians(x), degreesToRadians(y), degreesToRadians(z)
-		);
+			degreesToRadians(x), degreesToRadians(y), degreesToRadians(z));
 	}
 
 	Eigen::Quaternionf quaternionFromEulerAnglesDegrees(Eigen::Vector3f degrees)
@@ -29,12 +24,12 @@ namespace HOL
 
 	float degreesToRadians(float degrees)
 	{
-		return degrees * (M_PI / 180.0f);
+		return degrees * (std::numbers::pi_v<float> / 180.0f);
 	}
 
 	float radiansToDegrees(float radians)
 	{
-		return radians * (180.0f / M_PI);
+		return radians * (180.0f / std::numbers::pi_v<float>);
 	}
 
 	Eigen::Vector3f flipHandRotation(Eigen::Vector3f& rot)
