+++ HandOfLesserCommon/src/math/math_utils.h
@@ -2,12 +2,10 @@
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
-#include <d3d11.h>
-#include <openxr/openxr_platform.h>
-#include "openxr/openxr_structs.hpp"
 
 namespace HOL
 {
+
 	Eigen::Quaternionf quaternionFromEulerAngles(Eigen::Vector3f);
 	Eigen::Quaternionf quaternionFromEulerAnglesDegrees(Eigen::Vector3f degrees);
 	Eigen::Quaternionf quaternionFromEulerAnglesDegrees(float x, float y, float z);
@@ -17,4 +15,5 @@ namespace HOL
 
 	Eigen::Vector3f flipHandRotation(Eigen::Vector3f& rot);
 	Eigen::Vector3f flipHandTranslation(Eigen::Vector3f& trans);
+
 } // namespace HOL
